backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "ojoya backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.112.2",
    "firebase-admin>=6.5.0",
    "google-cloud-storage>=2.18.0",
    "httpx>=0.25.0",
    "langchain>=1.2.1",
    "langchain-google-genai>=4.1.3",
    "langgraph>=1.0.5",
    "Pillow>=10.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[project.scripts]
# 必要に応じてCLIコマンドを追加
# hello = "backend:hello"

[dependency-groups]
dev = [
    "ruff>=0.6.2",
    "fastapi-cli>=0.0.5",
    "pytest>=8.0.0",
]
